+++ class_person.rb
@@ -1,0 +1,21 @@
+class Person
+  attr_accessor :name, :age
+  attr_reader :id
+
+  def initialize(id, age, name = 'Unknown', parent_permission: true)
+    @id = id
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+  end
+
+  def of_age?
+    return true if @age >= 18
+  end
+
+  def can_use_services?
+    return true if of_age? || @parent_permission
+
+    false
+  end
+end
